DAOs now working. Updated spring version

+++ lacueva-control/src/main/java/com/lacueva/control/commons/Constants.java
@@ -2,21 +2,21 @@ package com.lacueva.control.commons;
 
 public class Constants {
 
-	public static final double DVD_PRICE = 15d;
-	public static final double BD_PRICE = 40d;
+	public static final float DVD_PRICE = 15f;
+	public static final float BD_PRICE = 40f;
 
 	/**
 	 * Weight in grams
 	 */
-	public static final double DVD_WEIGHT = 16.4d;
-	public static final double BD_WEIGHT = 16.4d;
-	public static final double ENVELOPE_WEIGHT = 1.56d;
-	public static final double ENVELOPE_FLAP_WEIGHT = 2.66d;
-	public static final double BULK10_WEIGHT = 40d;
-	public static final double BULK25_WEIGHT = 40d;
-	public static final double BULK50_WEIGHT = 41.25d;
-	public static final double BULK100_WEIGHT = 57d;
-	public static final double BOX_STD_WEIGHT = 1334d;
-	public static final double BOX_BIG_WEIGHT = 1d;
-	public static final double BOX_PAPERBOARD_WEIGHT = 1d;
+	public static final float DVD_WEIGHT = 16.4f;
+	public static final float BD_WEIGHT = 16.4f;
+	public static final float ENVELOPE_WEIGHT = 1.56f;
+	public static final float ENVELOPE_FLAP_WEIGHT = 2.66f;
+	public static final float BULK10_WEIGHT = 40f;
+	public static final float BULK25_WEIGHT = 40f;
+	public static final float BULK50_WEIGHT = 41.25f;
+	public static final float BULK100_WEIGHT = 57f;
+	public static final float BOX_STD_WEIGHT = 1334f;
+	public static final float BOX_BIG_WEIGHT = 1f;
+	public static final float BOX_PAPERBOARD_WEIGHT = 1f;
 }
